requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1603893126410/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1602530275854/work
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828496740/work
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
biopython @ file:///tmp/build/80754af9/biopython_1600298855952/work
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.12.5
cffi==1.14.0
chardet @ file:///tmp/build/80754af9/chardet_1605303175790/work
click==7.1.2
cryptography @ file:///tmp/build/80754af9/cryptography_1605544480695/work
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
faiss==1.6.3
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1604100403705/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
grpcio @ file:///tmp/build/80754af9/grpcio_1597424460877/work
h5py @ file:///tmp/build/80754af9/h5py_1593454121459/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
idna-ssl==1.1.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206602906/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
jedi @ file:///tmp/build/80754af9/jedi_1598371618777/work
Jinja2==2.11.2
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions @ file:///home/conda/feedstock_root/build_artifacts/jupyter_contrib_nbextensions_1602805459530/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1606148976284/work
jupyter-highlight-selected-word @ file:///home/conda/feedstock_root/build_artifacts/jupyter_highlight_selected_word_1603234228293/work
jupyter-latex-envs @ file:///home/conda/feedstock_root/build_artifacts/jupyter_latex_envs_1602788802079/work
jupyter-nbextensions-configurator @ file:///home/conda/feedstock_root/build_artifacts/jupyter_nbextensions_configurator_1602769491754/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras==2.3.1
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014557000/work
llvmlite==0.35.0
lxml @ file:///home/conda/feedstock_root/build_artifacts/lxml_1603104844695/work
Markdown @ file:///tmp/build/80754af9/markdown_1605111083232/work
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603373816490/work
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///tmp/build/80754af9/multidict_1600456401639/work
nb-conda==2.2.1
nb-conda-kernels @ file:///tmp/build/80754af9/nb_conda_kernels_1600445353977/work
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914804165/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
nglview @ file:///home/conda/feedstock_root/build_artifacts/nglview_1592924837339/work
notebook @ file:///tmp/build/80754af9/notebook_1601501581489/work
numba==0.52.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.4
pandas @ file:///tmp/build/80754af9/pandas_1602088135163/work
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120937332/work
parso==0.7.0
patsy==0.5.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1603822238230/work
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.13.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1605560620615/work/dist/ptyprocess-0.6.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyfasta==0.5.2
Pygments @ file:///tmp/build/80754af9/pygments_1604103097372/work
PyJWT==1.7.1
pynndescent==0.5.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141725711/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305763431/work
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1598379690588/work
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
seaborn==0.11.1
Send2Trash==1.5.0
six @ file:///tmp/build/80754af9/six_1605205335545/work
statsmodels==0.12.1
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets==4.3.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
umap-learn==0.5.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==0.90
xlrd==2.0.1
yarl @ file:///tmp/build/80754af9/yarl_1602693044244/work
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
